Refactor site navigation.
