+++ index.jl
@@ -1,0 +1,13 @@
+module Index
+using GenieFramework
+@genietools
+
+function ui()
+    [
+        a("Hill Equation", href = "/hill")
+    ]
+    
+end
+
+@page("/", ui)
+end
